added register state support, modified -l log behavior,
register state implemented and functional

modified the behavior so stf tracing can operate with or without
spike logging (-l --log=<fn>) .

memory record tracing was implemented previously, there are some
discrepencies that need to be investigated.

TODO:

added placeholders for vector register state. this is complicated
by the implementation of the vector register file. This is implemented
as a memory region, with variations for vlen, elen, slen.

Also need to add proper flag to stf header record to indicate that
the trace contains vector instructions. Spec says it is required
I do not know why.